refactor(health): move `health` model from `model` to `models` directory
